refactor: move unit tests from tests/Unit/ to app/Packages/ and update namespaces
