+++ src/app/Packages/Domains/Tests/User/Factory/SubscriptionFactoryTest.php
@@ -1,10 +1,9 @@
 <?php
 
-namespace Tests\Unit\Packages\Domains\User\Factory;
+namespace App\Packages\Domains\Tests\User\Factory;
 
 use App\Packages\Domains\User\Factory\SubscriptionFactory;
 use App\Packages\Domains\User\Subscription\Subscription;
-use App\Packages\Domains\User\Subscription\SubscriptionTier;
 use DateTimeImmutable;
 use PHPUnit\Framework\TestCase;
